fix(facade): isolate facade in own workspace + Cargo.lock (closes #198)
The facade contract's promise — stable contract id across releases —
was undefended by PR #205: facade was a workspace member, so any
workspace dep churn (dioxus, chrono, ml-dsa, freenet-stdlib) re-resolved
its transitive deps and rotated facade.wasm.

This commit moves the facade out of the workspace into its own
self-contained build with its own Cargo.lock.

Structure:

- contracts/facade-types/ — new tiny crate. Holds FacadePointer,
  FacadeMetadata, signed_payload, FACADE_MAX_PREV_APP_IDS. Only deps
  are ed25519-dalek + serde. Lives outside workspace so both consumers
  resolve against the same source.
- contracts/facade/ — now [workspace.exclude]'d. Self-contained
  Cargo.toml with =x.y.z pins on byteorder, ciborium, ed25519-dalek,
  freenet-stdlib. Path-deps on facade-types. Has its own Cargo.lock
  (committed) and its own target/ dir.
- common/src/facade.rs — collapsed to a re-export of facade-types so
  signer + UI keep their freenet-email-core::facade::* import paths.
- contracts/facade/src/lib.rs — imports from
  freenet_email_facade_types instead of freenet_email_core::facade.

Build pipeline:

- Makefile.toml build-facade now cd's into contracts/facade and runs
  cargo against the facade-local manifest. WASM output path moved to
  contracts/facade/target/.
- scripts/check-facade-byte-equal.sh updated to look at the new path.
- .github/workflows/build.yml: new step builds + lints + tests the
  facade in its own workspace (cargo test --workspace doesn't reach it).
- .github/workflows/check-contract-wasm.yml: path filter includes
  contracts/facade-types/.

Net effect: bumping any workspace dep cannot rotate facade.wasm,
because cargo resolves the facade build against
contracts/facade/Cargo.lock — not the workspace lockfile. Deliberate
facade dep bumps now require editing the =x.y.z pin in
contracts/facade/Cargo.toml, which is the only place where rotation
can be intentionally introduced.

Tests: 15 facade unit + 7 signer unit + 2 integration all pass.
Integration test (tools/web-container-sign/tests/facade_integration.rs)
still composes signer CLI output → on-chain validate_state across the
two workspace boundary, exercising both lockfiles end-to-end.

AGENTS.md §"Facade contract" updated to describe the new structure.

Closes #198.
